requirements.txt
asgiref==3.2.7
Django==3.0.6
gobject==0.1.0
pycairo==1.19.1
PyGObject==3.36.1
python-vlc==3.0.9113
pytz==2020.1
sqlparse==0.3.1
